# Conflicts:
#	src/main/webapp/WEB-INF/view/product/viewcategoryMessages.jsp
#	target/m2e-wtp/web-resources/META-INF/maven/com.foodmarket/foodmarket/pom.properties

+++ src/main/java/com/foodmarket/app/shopcar/controller/OrderItemController.java
@@ -43,7 +43,7 @@ public class OrderItemController {
 		System.out.println("saveRecord");
 		OrderRecord newOrderRecord = orderRecordService.save(orderRecord, userId);
 		
-		shopCartService.deleteByCustomerId(userId.intValue());
+		shopCartService.deleteByCustomerId(userId);
 		
 		return newOrderRecord;
 	}
